Fix secret contents
Currently the secret will not appear in the Mobile Control Panel,
the reason for this is that the Mobile Control Panel is expecting
the type of the service to be fh-sync-server and not fhsync.

The name field also needs to be updated to allow the icon for the
service to be discovered.

Finally, the name of the secret has been updated to match with the
other two values.