Update deprovision tasks to use ansible kubernetes modules
Currently our deprovision task uses the shell module to run the oc
tool to delete resources. This is messy and will fail if for some
reason the resource never existed in the first place.

This updates the deprovision task to use the ansible kubernetes
modules that are provided to the ansible task.

Also this change removes resources that are not currently deleted
in the current deprovision task.